Bug Fix
Fixed an issue where the json response contained values larger than
what the data type Int could hold

+++ app/src/main/java/ke/co/slick/researchapp/data/models/UsptoApiResponse.kt
@@ -31,5 +31,5 @@ data class Doc(
         val archiveUrl: String,
         val pdfPath: String,
         val year: String,
-        @Json(name = "_version_") val version: Int
+        @Json(name = "_version_") val version: Long
 )
